👷 Migrate from jCenter to Maven Central
Related: #noref
