+++ CHANGELOG.md
@@ -1,0 +1,11 @@
+# Changelog
+All notable changes to this project will be documented in this file.
+
+The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
+and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
+
+## [Unreleased]
+### Changed
+- Migration from jCenter to Maven Central 🎉
+- ‼️ Important ‼️ groupId has changed. New groupId is `io.github.ackeecz`.
+- Package name changed as well from `cz.ackee.ankoconstraintlayout` to `io.github.ackeecz.ankoconstraintlayout`
