+++ anko-constraint-layout/build.gradle
@@ -25,6 +25,7 @@ dependencies {
     implementation "org.jetbrains.kotlin:kotlin-stdlib:$kotlin_version"
     implementation 'androidx.appcompat:appcompat:1.0.2'
 
+    // TODO Change to implementation
     api 'androidx.constraintlayout:constraintlayout:1.1.3'
     api "org.jetbrains.anko:anko-sdk15:0.10.8"
 
@@ -34,13 +35,9 @@ dependencies {
     androidTestImplementation 'androidx.test.espresso:espresso-core:3.1.1'
 }
 
-repositories {
-    mavenCentral()
+ext {
+    PUBLISH_ARTIFACT_ID = 'anko-constraint-layout'
+    POM_DESCRIPTION = 'DSL implementation of constraint layout for anko'
 }
 
-//apply from: rootProject.file('gradle/gradle-mvn-push.gradle')
-
-apply from: rootProject.file('gradle/bintray/config.gradle')
-apply from: rootProject.file('gradle/bintray/install.gradle')
-apply from: rootProject.file('gradle/bintray/bintray.gradle')
-
+apply from: "${rootProject.projectDir}/gradle/mavencentral/publish.gradle"
