--- anko-constraint-layout/src/androidTest/java/cz/ackee/ankoconstraintlayout/ExampleInstrumentedTest.java
@@ -1,27 +1,0 @@
-package cz.ackee.ankoconstraintlayout;
-
-import android.content.Context;
-
-import org.junit.Test;
-import org.junit.runner.RunWith;
-
-import androidx.test.core.app.ApplicationProvider;
-import androidx.test.ext.junit.runners.AndroidJUnit4;
-
-import static org.junit.Assert.*;
-
-/**
- * Instrumented test, which will execute on an Android device.
- *
- * @see <a href="http://d.android.com/tools/testing">Testing documentation</a>
- */
-@RunWith(AndroidJUnit4.class)
-public class ExampleInstrumentedTest {
-    @Test
-    public void useAppContext() throws Exception {
-        // Context of the app under test.
-        Context appContext = ApplicationProvider.getApplicationContext();
-
-        assertEquals("cz.ackee.anko_constraint_layout.test", appContext.getPackageName());
-    }
-}
