+++ anko-constraint-layout/src/main/AndroidManifest.xml
@@ -1,1 +1,1 @@
-<manifest package="cz.ackee.anko_constraint_layout" />
+<manifest package="io.github.ackeecz.ankoconstraintlayout" />
