+++ app/build.gradle
@@ -27,7 +27,7 @@ dependencies {
     implementation 'androidx.appcompat:appcompat:1.0.2'
     implementation 'androidx.constraintlayout:constraintlayout:1.1.3'
 
-    implementation 'cz.ackee:anko-constraint-layout:1.2.0'
+    implementation 'io.github.ackeecz:anko-constraint-layout:2.0.0'
 
     testImplementation 'junit:junit:4.12'
     androidTestImplementation 'androidx.test:runner:1.1.1'
