+++ app/src/main/AndroidManifest.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="cz.ackee.anko_constraint_layout">
+    package="io.github.ackeecz.ankoconstraintlayout.sample">
 
     <application
         android:allowBackup="true"
@@ -10,7 +10,7 @@
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
 
-        <activity android:name="cz.ackee.ankoconstraintlayout.MainActivity">
+        <activity android:name="io.github.ackeecz.ankoconstraintlayout.sample.MainActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
