--- app/src/main/java/cz/ackee/ankoconstraintlayout/MainActivity.kt
@@ -1,17 +1,0 @@
-package cz.ackee.ankoconstraintlayout
-
-import android.os.Bundle
-import androidx.appcompat.app.AppCompatActivity
-import org.jetbrains.anko.setContentView
-
-/**
- * @author David Khol [david.khol@ackee.cz]
- * @since 18. 8. 2017
- **/
-class MainActivity : AppCompatActivity() {
-
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        MainActivityUI().setContentView(this)
-    }
-}
