+++ app/src/main/res/layout/activity_main.xml
@@ -5,7 +5,7 @@
 	xmlns:tools="http://schemas.android.com/tools"
 	android:layout_width="match_parent"
 	android:layout_height="match_parent"
-	tools:context="cz.ackee.ankoconstraintlayout.MainActivity">
+	tools:context="io.github.ackeecz.ankoconstraintlayout.sample.MainActivity">
 
     <TextView
         android:id="@+id/textView"
