+++ lib.properties
@@ -1,0 +1,14 @@
+VERSION_NAME=2.0.0
+VERSION_CODE=1
+GROUP=io.github.ackeecz
+SITE_URL=https://github.com/AckeeCZ/anko-constraint-layout
+GIT_URL=https://github.com/AckeeCZ/anko-constraint-layout.git
+POM_DEVELOPER_ID=ackee
+POM_DEVELOPER_NAME=Ackee
+POM_DEVELOPER_EMAIL=info@ackee.cz
+POM_LICENCE_NAME=The Apache Software License, Version 2.0
+POM_LICENCE_URL=http://www.apache.org/licenses/LICENSE-2.0.txt
+POM_SCM_CONNECTION=scm:git:github.com/AckeeCZ/anko-constraint-layout.git
+POM_SCM_DEVELOPER_CONNECTION=scm:git:ssh://github.com/AckeeCZ/anko-constraint-layout.git
+POM_SCM_URL=https://github.com/AckeeCZ/anko-constraint-layout/tree/main
+SONATYPE_URL=https://s01.oss.sonatype.org/service/local/staging/deploy/maven2/
